fix: do not try to list backups on emulator

+++ google-cloud-spanner/src/test/java/com/google/cloud/spanner/IntegrationTestEnv.java
@@ -19,6 +19,7 @@ package com.google.cloud.spanner;
 import static com.google.common.base.Preconditions.checkState;
 
 import com.google.api.gax.longrunning.OperationFuture;
+import com.google.cloud.spanner.testing.EmulatorSpannerHelper;
 import com.google.cloud.spanner.testing.RemoteSpannerHelper;
 import com.google.common.collect.Iterators;
 import com.google.spanner.admin.instance.v1.CreateInstanceMetadata;
@@ -165,17 +166,19 @@ public class IntegrationTestEnv extends ExternalResource {
       if (isOwnedInstance) {
         // Delete the instance, which implicitly drops all databases in it.
         try {
-          // Backups must be explicitly deleted before the instance may be deleted.
-          logger.log(
-              Level.FINE, "Deleting backups on test instance {0}", testHelper.getInstanceId());
-          for (Backup backup :
-              testHelper
-                  .getClient()
-                  .getDatabaseAdminClient()
-                  .listBackups(testHelper.getInstanceId().getInstance())
-                  .iterateAll()) {
-            logger.log(Level.FINE, "Deleting backup {0}", backup.getId());
-            backup.delete();
+          if (!EmulatorSpannerHelper.isUsingEmulator()) {
+            // Backups must be explicitly deleted before the instance may be deleted.
+            logger.log(
+                Level.FINE, "Deleting backups on test instance {0}", testHelper.getInstanceId());
+            for (Backup backup :
+                testHelper
+                    .getClient()
+                    .getDatabaseAdminClient()
+                    .listBackups(testHelper.getInstanceId().getInstance())
+                    .iterateAll()) {
+              logger.log(Level.FINE, "Deleting backup {0}", backup.getId());
+              backup.delete();
+            }
           }
           logger.log(Level.FINE, "Deleting test instance {0}", testHelper.getInstanceId());
           instanceAdminClient.deleteInstance(testHelper.getInstanceId().getInstance());
